Docker file base version
This is the base version of the docker file, which enables you to build titan with a custom tinkerpop version "3.0.0-MY-SNAPSHOT", you may wish to change this to the appropriate version as you wish. 

This is configured to build titan with TITAN_VERSION="0.9.0-SNAPSHOT" and from the latest github branch: TITAN_BRANCH="titan09"

This doesn't include installation of the dependencies cassandra / elasticsearch. It is expected that you would install them as separate docker containers.